+++ packages/core/src/lib/store.types.ts
@@ -1,13 +1,15 @@
 import { BehaviorSubject } from "rxjs";
 
-import { WalletModule } from "./wallet/wallet.types";
+import { WalletModule } from "./wallet";
+
+export type ModuleState = WalletModule;
 
 export interface AccountState {
   accountId: string;
 }
 
 export interface WalletSelectorState {
-  modules: Array<WalletModule>;
+  modules: Array<ModuleState>;
   accounts: Array<AccountState>;
   selectedWalletId: string | null;
 }
@@ -16,7 +18,7 @@ export type WalletSelectorAction =
   | {
       type: "SETUP_WALLET_MODULES";
       payload: {
-        modules: Array<WalletModule>;
+        modules: Array<ModuleState>;
         accounts: Array<AccountState>;
         selectedWalletId: string | null;
       };
